# Faster as_u128 and to_u128_le